move validity check to job
